fix!: remove old ingest client
