+++ indykite_sdk_node/src/index.ts
@@ -3,8 +3,7 @@ export * from './sdk/authorization';
 export { IdentityClient } from './sdk/identity';
 export { ConfigClient } from './sdk/config';
 export { HTTPClient, HTTPResponse } from './sdk/http';
-export { IngestClient } from './sdk/ingest';
-export * from './sdk/ingest_v2';
+export * from './sdk/ingest';
 export * from './sdk/model';
 export {
   ChangePasswordRequest,
